+++ electron/mpris.js
@@ -1,34 +1,38 @@
 var Player = require('mpris-service');
-
 var player = Player({
-	name: 'nodejs',
-	identity: 'Node.js media player',
+	name: 'Petal',
+	identity: 'douban.FM player',
 	supportedUriSchemes: ['file'],
 	supportedMimeTypes: ['audio/mpeg', 'application/ogg'],
 	supportedInterfaces: ['player']
 });
 
+player.getPosition = function() {
+  // return the position of your player
+  return 0;
+}
+
+// Events
+var events = ['raise', 'quit', 'next', 'previous', 'pause', 'playpause', 'stop', 'play', 'seek', 'position', 'open', 'volume', 'loopStatus', 'shuffle'];
+events.forEach(function (eventName) {
+	player.on(eventName, function () {
+		console.log('Event:', eventName, arguments);
+	});
+});
+
 player.on('quit', function () {
 	process.exit();
 });
-export function setMprisState({title, album, artist}) {
+
+exports = function setMprisMetadata(song) {
     player.metadata = {
 		'mpris:trackid': player.objectPath('track/0'),
 		'mpris:length': 60 * 1000 * 1000, // In microseconds
-		'mpris:artUrl': 'http://www.adele.tv/images/facebook/adele.jpg',
-		'xesam:title': title,
-		'xesam:album': album,
-		'xesam:artist': artist
+		'mpris:artUrl': song.picture,
+		'xesam:title': song.title,
+		'xesam:album': song.album,
+		'xesam:artist': [song.artist]
     }
     player.playbackStatus = Player.PLAYBACK_STATUS_PLAYING;
-    console.log(`Now playing: ${title} - ${artist} - ${album}`);
+    console.log(`Now playing: ${song.title} - ${song.artist} - ${song.album}`);
 }
-
-setTimeout( ()=> {
-    const state = {
-        title:'Love Story',
-        album:'Secret',
-        artist:['Taylor Swift']
-    };
-    setMprisState(state);
-},1000);
