+++ src/components/FM/Cover/index.jsx
@@ -14,8 +14,6 @@ import {
   playtimeSet,
   updateSidSsid
 } from '../../../actions/fm/actions'
-//import {setMprisMetadata} from './mpris'
-
 class Cover extends Component {
   constructor(props) {
     super(props)
@@ -125,7 +123,6 @@ class Cover extends Component {
    * @memberof Cover
    */
   setCover = (song, pattern, openWithPlaying = true) => {
-//    setMprisMetadata(song);
     this.setState(
       {
         playing: openWithPlaying ? true : false,
