+++ webpack/config.electron.js
@@ -15,5 +15,10 @@ module.exports = {
   target: 'electron-main',
   node: {
     __dirname: false
+  },
+  resolve: {
+    alias: {
+      jsbi: path.resolve(__dirname, '..', 'node_modules', 'jsbi', 'dist', 'jsbi-cjs.js')
+    }
   }
 }
